+++ dataland-e2etests/src/test/kotlin/org/dataland/e2etests/Constants.kt
@@ -13,4 +13,4 @@ const val ADMIN_USER_PASSWORD = "test"
 const val SOME_USER_NAME = "some_user"
 const val SOME_USER_PASSWORD = "test"
 
-const val TEASER_COMPANY_INDEX_IN_FIXTURES = 2
+val TEASER_COMPANY_INDEX_IN_FIXTURES: Int = System.getenv("TEASER_COMPANY_INDEX_IN_FIXTURES").toInt()
