+++ environments/.env.template
@@ -16,6 +16,9 @@ SKYMINDER_URL="${SKYMINDER_URL}"
 SKYMINDER_PW="${SKYMINDER_PW}"
 SKYMINDER_USER="${SKYMINDER_USER}"
 
+#Variable to define a teaser company which is available together with its data sets without authorization via token
+TEASER_COMPANY_INDEX_IN_FIXTURES=${TEASER_COMPANY_INDEX_IN_FIXTURES}
+
 #Variables only for deployment script
 BACKEND_DOCKERFILE="DockerfileBackend"
 FRONTEND_DOCKERFILE="DockerfileFrontend"
